+++ text.cpp
@@ -6,7 +6,6 @@
 #include <vector>
 #include "text.hpp"
 #include "fpfmt/fpfmt.hpp"
-#include "fpfmt/fpfmt.h"
 #include "errors.hpp"
 
 /**
@@ -174,17 +173,6 @@ std::string format_commandline(int argc, char **argv) {
 	}
 
 	return out;
-}
-
-// for jsonpull to call from C
-char *dtoa_shortest(double val) {
-	std::string s = fpfmt::dtoa(val);
-	char *dup = strdup(s.c_str());
-	if (dup == NULL) {
-		perror("strdup");
-		exit(EXIT_MEMORY);
-	}
-	return dup;
 }
 
 // to work with data from https://github.com/kmike/text-unidecode